move and rename deepPartialSchema -> makeConventionDeepPartialSchema
