+++ shared/src/zodUtils.ts
@@ -121,55 +121,3 @@ export type ZodSchemaWithInputMatchingOutput<T> = z.ZodType<
   T,
   z.core.$ZodTypeInternals<T, T>
 >;
-
-export const deepPartialSchema = (schema: z.ZodTypeAny): z.ZodTypeAny => {
-  if (schema.def.type === "object") {
-    const newShape: Record<string, z.ZodTypeAny> = {};
-    for (const [key, value] of Object.entries(
-      (schema as unknown as z.ZodObject<any>).shape,
-    )) {
-      newShape[key] = deepPartialSchema(value as z.ZodTypeAny).optional();
-    }
-    return z.object(newShape);
-  }
-
-  if (schema.def.type === "array") {
-    return z.array(
-      deepPartialSchema((schema as unknown as z.ZodArray<any>).element),
-    );
-  }
-
-  if (schema.def.type === "optional") {
-    return deepPartialSchema(
-      (schema as unknown as z.ZodOptional<any>).unwrap(),
-    ).optional();
-  }
-
-  if (schema.def.type === "nullable") {
-    return deepPartialSchema(
-      (schema as unknown as z.ZodNullable<any>).unwrap(),
-    ).nullable();
-  }
-
-  if (schema.def.type === "union") {
-    const unionSchema = schema as unknown as z.ZodUnion<any>;
-    return z.union(
-      unionSchema.options.map((option: z.ZodTypeAny) =>
-        deepPartialSchema(option),
-      ),
-    );
-  }
-
-  if (schema.def.type === "intersection") {
-    const def = schema.def as unknown as {
-      left: z.ZodTypeAny;
-      right: z.ZodTypeAny;
-    };
-    return z.intersection(
-      deepPartialSchema(def.left),
-      deepPartialSchema(def.right),
-    );
-  }
-
-  return schema;
-};
